pyproject.toml
[tool.poetry]
name = "trans-manager"
version = "0.1"
description = "managering to tranlation work"
authors = ["bh224 <ericakim224@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "^4.1.7"
django-environ = "^0.9.0"
mysqlclient = "^2.1.1"
djangorestframework = "^3.14.0"
openpyxl = "^3.1.1"
boto3 = "^1.26.73"
django-debug-toolbar = "^3.8.1"
django-cors-headers = "^3.14.0"
requests = "^2.28.2"
Scrapy = "^2.8.0"
celery = "^5.2.7"
redis = "^4.5.1"
django-celery-results = "^2.4.0"
django-celery-beat = "^2.4.0"
scrapy-djangoitem = "^1.1.1"

[tool.poetry.dev-dependencies]
black = {version = "^23.1.0", allow-prereleases = true}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
